changed datatype for classification grid to np.uint8, since before it just used int which is actually resulting in int64 and therefore uses much more memory

+++ occpy/OccPy.py
@@ -610,7 +610,7 @@ class OccPy:
         # Create Classification grid
         self.logger.info("Classify Grid")
         tic = time.time()
-        self.Classification = np.zeros((self.grid_dim['nx'], self.grid_dim['ny'], self.grid_dim['nz']), dtype=int)
+        self.Classification = np.zeros((self.grid_dim['nx'], self.grid_dim['ny'], self.grid_dim['nz']), dtype=np.uint8)
 
         self.Classification[np.logical_and.reduce((self.Nhit > 0, self.Nmiss >= 0, self.Nocc >= 0))] = 1  # voxels that were observed
         self.Classification[np.logical_and.reduce((self.Nhit == 0, self.Nmiss > 0, self.Nocc >= 0))] = 2  # voxels that are empty
